fastapi-async-api/requirements.txt
fastapi
uvicorn
pillow
aiofiles
aioboto3
python-multipart